refactor(ui): split the context-window control out of the model picker
chat-model-picker.ts crossed the 700-line lint cap after the switch moved
in; the row renderer and its params now live in chat-context-window-control.ts.
